Update dimacs to latest commit to fix deprecation warning.
Dimacs depended on a too old bitflags version which produced a
deprecation warning. In the git repo of dimacs this is already fixed,
but there is no release with the fix, resulting in cargo not updating to
the fixed version. The workaround is to specify the git repo in the
Cargo.toml.